pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "igelfs"
description = "Python implementation of the IGEL filesystem."
version = "2.4.1"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
  {name = "Zack Didcott"}
]
maintainers = [
  {name = "Zack Didcott"}
]
dependencies = [
  "pillow",
  "python-magic",
  "rsa",
]
requires-python = ">= 3.12"

[project.optional-dependencies]
convert = ["pyparted"]
tests = ["pytest"]

[project.urls]
Homepage = "https://github.com/Zedeldi/igelfs"

[project.scripts]
igelfs-cli = "igelfs.cli:main"

[tool.isort]
profile = "black"

[tool.mypy]
disable_error_code = [
  "import-not-found",
  "import-untyped",
]
enable_error_code = [
  "ignore-without-code"
]
show_error_codes = true
warn_unused_ignores = true

[tool.pytest.ini_options]
markers = [
  "inf: tests require LXOS configuration file",
  "slow: tests take a while to complete",
]
